requirements.txt
selenium
google-cloud-vision
openai
google-generativeai
PyMuPDF
Pillow
tqdm
tiktoken
flask
selenium  
google-cloud-vision  
openai  
google-generativeai  
PyMuPDF  
Pillow  
pyyaml  
tqdm  
requests  
tiktoken 
